Fix scoping of on-disk state to reflect system type/scope/id
Previously we only considered API key to determine which state to save,
which meant that in configurations with multiple servers but a single
organization API key we'd usually loose statistics on restarts, or
get the wrong values for diff-ing the query statistics.

This is a backwards-incompatible change for the on-disk format and
therefore bumps the version from 1 to 2. Effectively this will show
as one period of no data after upgrading to this version, as the
previously saved counter values in the state file won't be used.